feat: enable componentized browser js
